Add -use_var_is_introduced option
This fixes a possible regression from the MiniZinc 2015 challenge
results.  Previous versions of Chuffed used the "output_var" annotation
to determine whether a variable was introduced or not.  I changed that
to use the "var_is_introduced" annotation instead, but this made solving
slower on some problems.  With the new -use_var_is_introduced option
true, the "var_is_introduced" is used; with it false (the default),
"output_var" is used as before.